+++ sql/samples/core-data.sql
@@ -7,12 +7,12 @@
 -- Started 2023-08-09
 -- ---------------------------------------------------------------------
 
-INSERT INTO dataset (id, code, hdx_stub, title, provider_code, provider_name)
+INSERT INTO dataset (id, hdx_id, hdx_stub, title, provider_code, provider_name)
 VALUES
 (1, 'c3f001fa-b45b-464c-9460-1ca79fd39b40', 'dataset01', 'Dataset #1', 'provider01', 'Provider #1'),
 (2, '7cf3cec8-dbbc-4c96-9762-1464cd0bff75', 'dataset02', 'Dataset #2', 'provider02', 'Provider #2');
 
-INSERT INTO resource (id, dataset_ref, code, filename, format, update_date, is_hxl)
+INSERT INTO resource (id, dataset_ref, hdx_id, filename, format, update_date, is_hxl)
 VALUES
 (1, 1, '90deb235-1bf5-4bae-b231-3393222c2d01', 'resource-01.csv', 'csv', '2023-06-01 00:00:00', TRUE),
 (2, 1, 'b9e438e0-b68a-49f9-b9a9-68c0f3e93604', 'resource-02.xlsx', 'xlsx', '2023-07-01 00:00:00', TRUE),